Added: Created all connections for projects and modified create and update calls to allow for creation of atributes in said connections

+++ src/controller/projectController.ts
@@ -1,6 +1,9 @@
 import { type RequestHandler, type Request, type Response } from "express";
 import { Project } from "../models/project/projects";
 import { Position } from "../models/position/positions";
+import { ExpirationDateProject } from "../models/project/expiration_date_project";
+import { ClosedProject } from "../models/project/closed_project";
+import { Client } from "../models/client/clients";
 
 export const createProject: RequestHandler = async (
   req: Request,
@@ -14,22 +17,113 @@ export const createProject: RequestHandler = async (
     });
   }
 
-  Project.create(
-    { ...req.body })
-    .then((data: Project) => {
-      return res.status(201).json({
-        status: "Success",
-        message: "Project created successfully",
-        payload: data,
+  const { has_expiration_date, expiration_date, project_title, project_description, start_date, general_status, closed_status, closed_reason, client_name } = req.body;
+
+  if (!project_title || !project_description || !start_date || !general_status || !client_name || typeof has_expiration_date !== 'boolean') {
+    return res.status(400).json({
+        status: 'error',
+        message: 'All fields are required',
+        payload: null
+    });
+  }
+
+  const client = await Client.findOne({
+    where: {
+      client_name
+    }
+  })
+  
+  if (!client) {
+    return res.status(404).json({
+      status: 'error',
+      message: 'Client not found',
+      payload: null
+    });
+  }
+
+  const client_id = client.id;
+
+  if(!["In Preparation", "Active", "Closed"].includes(general_status)) {
+    return res.status(400).json({ 
+        status: 'error',
+        message: 'Invalid general status provided',
+        payload: null
+    });
+  }
+
+  if (has_expiration_date && !expiration_date) {
+      return res.status(400).json({
+        status: 'error',
+        message: 'Expiration date not provided',
+        payload: null
       });
-    })
-    .catch((error: Error) => {
+  }
+
+  if(general_status == "Closed"){
+    if(!closed_status || !closed_reason) {
+      return res.status(400).json({
+        status: 'error',
+        message: 'Closed status or reason not provided',
+        payload: null
+      });
+    }
+
+    if(!["Completed", "Cancelled"].includes(closed_status)) {
+      return res.status(400).json({ 
+          status: 'error',
+          message: 'Invalid closed status provided',
+          payload: null
+      });
+    }
+  
+  }
+
+  try{
+
+    const project = await Project.create({ ...req.body, client_id });
+
+    if (has_expiration_date) {
+      await ExpirationDateProject.create({
+        project_id: project.id,
+        expiration_date,
+      })
+      .catch((error: Error) => {
+        return res.status(500).json({
+          status: "Error",
+          message: "There was an error creating the expiration date",
+          payload: error.message,
+        });
+      });
+    }
+
+    if(general_status == "Closed"){
+      await ClosedProject.create({
+        project_id: project.id,
+        closed_status,
+        closed_reason
+      })
+      .catch((error: Error) => {
+        return res.status(500).json({
+          status: "Error",
+          message: "There was an error creating the closed project",
+          payload: error.message,
+        });
+      });
+    }
+
+    return res.status(201).json({
+      status: "Success",
+      message: "Project created successfully",
+      payload: project,
+    });
+  }
+    catch (error: any)  {
       return res.status(500).json({
         status: "Error",
         message: "Project not created",
         payload: error.message,
       });
-    });
+    };
 };
 
 export const getProjects: RequestHandler = async (
@@ -70,26 +164,154 @@ export const updateProject: RequestHandler = async (
   }
 
   const id = req.params.id;
-  Project.update({ ...req.body }, { where: { id } })
-    .then((isUpdated) => {
-      if (isUpdated) {
-        return res.status(200).json({
-          status: "Success",
-          message: "Project updated successfully",
-          payload: { ...req.body },
-        });
-      }
-      return res.status(500).json({
-        status: "Success",
-        message: "Something happened updating the product",
-        payload: null,
+
+  //Validations
+  const { has_expiration_date, expiration_date, project_title, project_description, start_date, general_status, closed_status, closed_reason, client_name } = req.body;
+
+  if (!project_title || !project_description || !start_date || !general_status || !client_name || typeof has_expiration_date !== 'boolean') {
+    return res.status(400).json({
+        status: 'error',
+        message: 'All fields are required',
+        payload: null
+    });
+  }
+
+  const client = await Client.findOne({
+    where: {
+      client_name
+    }
+  })
+  
+  if (!client) {
+    return res.status(404).json({
+      status: 'error',
+      message: 'Client not found',
+      payload: null
+    });
+  }
+
+  const client_id = client.id;
+
+  if(!["In Preparation", "Active", "Closed"].includes(general_status)) {
+    return res.status(400).json({ 
+        status: 'error',
+        message: 'Invalid general status provided',
+        payload: null
+    });
+  }
+
+  if (has_expiration_date && !expiration_date) {
+      return res.status(400).json({
+        status: 'error',
+        message: 'Expiration date not provided',
+        payload: null
       });
+  }
+
+  if(general_status == "Closed"){
+    if(!closed_status || !closed_reason) {
+      return res.status(400).json({
+        status: 'error',
+        message: 'Closed status or reason not provided',
+        payload: null
+      });
+    }
+
+    if(!["Completed", "Cancelled"].includes(closed_status)) {
+      return res.status(400).json({ 
+          status: 'error',
+          message: 'Invalid closed status provided',
+          payload: null
+      });
+    }
+  
+  }
+
+  //check project exists
+  Project.findByPk(id)
+    .then((data: Project | null) => {
+      if (data) {
+
+        if (data.general_status === "Closed") {
+          return res.status(400).json({
+            status: "error",
+            message: "Project status is 'Cancelled' and cannot be updated",
+            payload: null
+          });
+        }
+
+        if(has_expiration_date){
+          ExpirationDateProject.findOne({
+            where: {
+              project_id: id
+            }
+          })
+          .then((prevExpDate: ExpirationDateProject | null) => {
+
+            if(prevExpDate){
+              prevExpDate.update({expiration_date})   //if there is a previous expiration date update it
+              .then((isUpdatedExpiration) => {
+
+                if(!isUpdatedExpiration){
+                  return res.status(500).json({
+                    status: "Success",
+                    message: "Something happened updating the project (Expiration Date)",
+                    payload: null,
+                  });
+                }
+              });
+            } else{
+                ExpirationDateProject.create( {   //if there isn't one create it
+                  project_id: Number(id),
+                  expiration_date,
+                })
+            }
+          });
+          
+        } else{
+          ExpirationDateProject.destroy({    //if has_expiration_date is false destroy associated expiration date if any
+            where: {
+              project_id: Number(id)
+            }
+          });
+        }
+
+        Project.update({ ...req.body }, { where: { id } })
+          .then((isUpdated) => {
+            if (isUpdated) {
+              return res.status(200).json({
+                status: "Success",
+                message: "Project updated successfully",
+                payload: { ...req.body },
+              });
+            }
+            return res.status(500).json({
+              status: "Success",
+              message: "Something happened updating the project",
+              payload: null,
+            });
+          })
+          .catch((error: Error) => {
+            return res.status(500).json({
+              status: "Error",
+              message: `Project not updated: ${error.message}`,
+              payload: null,
+            });
+          });
+
+      } else{
+        return res.status(404).json({
+          status: 'error',
+          message: 'Project not found',
+          payload: null
+      });
+      }
     })
     .catch((error: Error) => {
       return res.status(500).json({
         status: "Error",
-        message: `Project not updated: ${error.message}`,
-        payload: null,
+        message: "Error updating Project",
+        payload: error.message,
       });
     });
 };
@@ -99,28 +321,41 @@ export const deleteProject: RequestHandler = async (
   res: Response
 ) => {
   const { id } = req.body;
-  Project.destroy({ where: { id } })
-    .then((isDeleted) => {
-      if (isDeleted) {
-        return res.status(200).json({
-          status: "Success",
-          message: "Project deleted successfully",
-          payload: { ...req.body },
+
+  //Make sure the project exists
+  Project.findByPk(id)
+  .then((data: Project | null) => {
+      if (data) {
+        Project.destroy({ where: { id } })
+          .then((isDeleted) => {
+            if (isDeleted) {
+              return res.status(200).json({
+                status: "Success",
+                message: "Project deleted successfully",
+                payload: { ...req.body },
+              });
+            }
+            return res.status(500).json({
+              status: "Error",
+              message: "Project not deleted",
+              payload: null,
+            });
+          })
+      } else {
+        return res.status(404).json({
+            status: 'error',
+            message: 'Project not found',
+            payload: null
         });
       }
-      return res.status(500).json({
-        status: "Error",
-        message: "Project not deleted",
-        payload: null,
-      });
-    })
-    .catch((error: Error) => {
-      return res.status(500).json({
-        status: "Error",
-        message: "Error deleting Project",
-        payload: error.message,
-      });
+  })
+  .catch((error: Error) => {
+    return res.status(500).json({
+      status: "Error",
+      message: "Error deleting Project",
+      payload: error.message,
     });
+  });
 };
 
 export const getProjectById: RequestHandler = async (
@@ -135,11 +370,19 @@ export const getProjectById: RequestHandler = async (
       }
     })
     .then((data: Project | null) => {
-      return res.status(200).json({
-        status: "Success",
-        message: "Project retrieved successfully",
-        payload: data,
-      });
+      if (data) {
+        return res.status(200).json({
+          status: "Success",
+          message: "Project retrieved successfully",
+          payload: data,
+        });
+      } else {
+          return res.status(404).json({
+            status: 'error',
+            message: 'Project not found',
+            payload: null
+        });
+      }
     })
     .catch((error: Error) => {
       return res.status(500).json({
